==> Lessons/18_CraftingReports.Rmd <==

---
title: "17: Crafting Reports"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2019"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## LESSON OBJECTIVES
1. Describe the purpose of using R Markdown as a communication and workflow tool
2. Incorporate Markdown syntax into documents
3. Communicate the process and findings of an analysis session in the style of a report

## BASIC R MARKDOWN DOCUMENT STRUCTURE

1. **YAML Header** surrounded by --- on top and bottom
    + YAML templates include options for html, pdf, word, markdown, and interactive
    + More information on formatting the YAML header can be found in the cheat sheet
2. **R Code Chunks** surrounded by ``` on top and bottom
    + Create using `Cmd/Ctrl` + `Alt` + `I`
    + Can be named {r name} to facilitate navigation and autoreferencing
    + Chunk options allow for flexibility when the code runs and when the document is knitted
3. **Text** with formatting options for readability in knitted document

A handy cheat sheet for R markdown can be found [here](https://www.rstudio.com/wp-content/uploads/2015/03/rmarkdown-reference.pdf). Another one can be found [here](https://www.rstudio.com/wp-content/uploads/2015/02/rmarkdown-cheatsheet.pdf).

## WHY R MARKDOWN?

<Fill in our discussion below with bullet points. Use italics and bold for emphasis (hint: use the cheat sheets to figure out how to make bold and italic text).>

* *italicize* using asterisks surrounding a word
* **bold** using two asterisks on each side of a word or phrase
* Documentation of what you are doing
* Allows for text, graphics, and code all in one clean(ish) document
* Control over what is displayed
* Easy to share/distribute
* Allows for more organization

## TEXT EDITING CHALLENGE

Create a table below that details the example datasets we have been using in class. The first column should contain the names of the datasets and the second column should include some relevant information about the datasets. (Hint: use the cheat sheets to figure out how to make a table in Rmd)

```{r table creation}
```
Dataset Name  | Description
--------------|-------------
EPA Air       | Air quality data
Leaf Litter   | Litter collection data
LTER Lake Data| Lake data from NTL


## R CHUNK EDITING CHALLENGE
### Installing packages
Create an R chunk below that installs the package `knitr`. Instead of commenting out the code, customize the chunk options such that the code is not evaluated (i.e., not run). 

```{r install package, eval = FALSE}
install.packages("knitr")
```


### Setup
Create an R chunk below called "setup" that checks your working directory, loads the packages `tidyverse` and `knitr`, and sets a ggplot theme. Remember that you need to disable R throwing a message, which containts a check mark that cannot be knitted.

Load the NTL-LTER_Lake_Nutrients_Raw dataset, display the head of the dataset, and set the date column to a date format.

Customize the chunk options such that the code is run but is not displayed in the final document.

```{r setup, message = FALSE, echo = FALSE}
getwd()
library(tidyverse)
library(knitr)
library(viridis)

# Set theme
mytheme <- theme_classic() +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)

NTL.Nutrients <- read.csv("./Data/Raw/NTL-LTER_Lake_Nutrients_Raw.csv")

head(NTL.Nutrients)

class(NTL.Nutrients$sampledate)
NTL.Nutrients$sampledate <- as.Date(NTL.Nutrients$sampledate, "%m/%d/%y")
class(NTL.Nutrients$sampledate)
```


### Data Exploration, Wrangling, and Visualization
Create an R chunk below to create a processed dataset do the following operations: 

* Include all columns except lakeid, depth_id, and comments
* Include only surface samples (depth = 0 m)

```{r Data Processing}
NTL.Nutrients.Processed <- NTL.Nutrients %>%
  select(-c("lakeid", "depth_id", "comments")) %>%
  filter(depth == 0)

```


Create a second R chunk to create a summary dataset with the mean, minimum, maximum, and standard deviation of total nitrogen concentrations for each lake. Create a second summary dataset that is identical except that it evaluates total phosphorus. Customize the chunk options such that the code is run but not displayed in the final document. 

```{r Summary, echo = FALSE}
NTL.Nutrients.Nitrogen.Summary <- NTL.Nutrients.Processed %>%
  filter(!is.na(tn_ug)) %>%
  group_by(lakename) %>%
  summarise(meanN = mean(tn_ug), minN = min(tn_ug), maxN = max(tn_ug), stdevN = sd(tn_ug))

NTL.Nutrients.Phosphorus.Summary <- NTL.Nutrients.Processed %>%
  filter(!is.na(tp_ug)) %>%
  group_by(lakename) %>%
  summarise(meanP = mean(tp_ug), minP = min(tp_ug), maxP = max(tp_ug), stdevP = sd(tp_ug))

```


Create a third R chunk that uses the function `kable` in the knitr package to display two tables: one for the summary dataframe for total N and one for the summary dataframe of total P. Use the `caption = " "` code within that function to title your tables. Customize the chunk options such that the final table is displayed but not the code used to generate the table. 

```{r kable, echo = FALSE}
Nitrogen.Table <- kable(NTL.Nutrients.Nitrogen.Summary, caption = "Nitrogen Summary")
Phos.Table <- kable(NTL.Nutrients.Phosphorus.Summary, caption = "Nitrogen Summary")

Nitrogen.Table
Phos.Table
```


Create a fourth and fifth R chunk that generates two plots (one in each chunk): one for total N over time with different colors for each lake, and one with the same setup but for total P. Decide which geom option will be appropriate for your purpose, and select a color palette that is visually pleasing and accessible. Customize the chunk options such that the final figures are displayed but not the code used to generate the figures. In addition, customize the chunk options such that the figures are aligned on the left side of the page. Lastly, add a fig.cap chunk option to add a caption (title) to your plot that will display underneath the figure.

```{r N Plot, echo = FALSE, fig.align = "left", fig.cap = "Total Nitrogen Over Time"}
ggplot() +
  geom_line(data = NTL.Nutrients.Processed, aes(x = sampledate, tn_ug, color = lakename)) +
  scale_color_viridis_d(option = "inferno", begin = 0.1, end = 0.8) +
  labs(x = "Sample Date", y = "Total Nitrogen (ug/L)", color = "")
  
```

```{r P Plot, echo = FALSE, fig.align = "left", fig.cap = "Total Phosphorus Over Time"}
ggplot() +
  geom_line(data = NTL.Nutrients.Processed, aes(x = sampledate, tp_ug, color = lakename)) +
  scale_color_viridis_d(option = "inferno", begin = 0.1, end = 0.8) +
  labs(x = "Sample Date", y = "Total Phosphorus (ug/L)", color = "")
  
```


### Other options
What are the chunk options that will suppress the display of errors, warnings, and messages in the final document? 

> ANSWER: Respectively: error, warning, message

### Communicating results
Write a paragraph describing your findings from the R coding challenge above. This should be geared toward an educated audience but one that is not necessarily familiar with the dataset. Then insert a horizontal rule below the paragraph. Below the horizontal rule, write another paragraph describing the next steps you might take in analyzing this dataset. What questions might you be able to answer, and what analyses would you conduct to answer those questions?

This is paragraph one, describing the data set. The data has various data about lake conditions such as total nitrogen, total phosphorous, and species of each chemical. Other data including temperature and sampling conditions are also included.

***

This is the second paragraph, describing the next steps I would take to analyze this data. It would be useful to have a specific research question, so that is probably where I would go next. Doing some statisitical tests across the lakes or across time seem like good starting places for more exploration as well. 

## KNIT YOUR PDF
When you have completed the above steps, try knitting your PDF to see if all of the formatting options you specified turned out as planned. This may take some troubleshooting. 

## OTHER R MARKDOWN CUSTOMIZATION OPTIONS
We have covered the basics in class today, but R Markdown offers many customization options. A word of caution: customizing templates will often require more interaction with LaTeX and installations on your computer, so be ready to troubleshoot issues.

Customization options for pdf output include: 

* Table of contents
* Number sections
* Control default size of figures
* Citations
* Template (more info [here](http://jianghao.wang/post/2017-12-08-rmarkdown-templates/))

pdf_document:  
    toc: true  
    number_sections: true  
    fig_height: 3  
    fig_width: 4  
    citation_package: natbib  
    template:
